Remove Loader from node context
